redo emacs config
the old config is now gone as it was a pile of junk.
new ~/.emacs is now seperated into sections which are denoted with ;; style comments.
i now use lsp, helm and company mode.
things still need to be improved, mainly figuring if the use-package order should be more rigous in what depends on what.
i provide a makefile for installing emacs onto a new system.